Moved ViewModel to correct package
