+++ src/main/java/fworks/views/UpdateCourseMembershipScreen.java
@@ -1,6 +1,7 @@
 package fworks.views;
 
 import ia.controllers.UpdateCourseMembershipController;
+import ia.viewmodels.CourseMembershipViewModel;
 
 import javax.swing.*;
 import java.awt.*;
